requirements.txt
python-fasthtml>=0.12.33
faststrap>=0.6.1
python-dotenv>=1.0.1
